backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djoser==2.3.1
django-filter==25.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
pybase62==1.0.0
python-dotenv==1.1.1
Pillow==11.3.0
requests==2.32.4
